requirements.txt
pytube
scrapetube
langchain-openai
tiktoken
langchain
python-telegram-bot
asyncio
youtube-transcript-api
